Bitmaps: add bitmap mode
This new mode work in the same way as the existing font mode: It allows drawing
solid (both white and black pixels in the bitmap affect the output) or
transparent (only white pixels in the bitmap affect the output). Transparent
bitmaps may be used to draw multiple bitmaps over each other, for example when
drawing animations. As a side effect, transparent XOR mode now also works by
setting the bitmap mode to transparent and the draw color to 2.

Before this commit, the bitmaps XBM bitmaps were being drawn as solid
object, while the regular bitmaps were drawn as transparent.

The structure of the code is based on how `u8g2_font_decode_len()` handles the
same thing when drawing glyphs.