Issue #340, add cimg_annotate_screenshot example
CImg is a lightweight, header-only image processing library. The
makefile automatically makes a git clone of this library when building.

This example shows how to capture a screenshot using the bitmap device,
convert it to a CImg bitmap, scale it, add (in a higher resolution than
the original screenshot) annotations and write out a PNG file.

This is a reduced version of a bigger program, so it might be a bit more
complicated that needed just for the example.