treewide: Fix address mode setup on SSD1306-like displays
Previously the addressing mode was set as 0x00. However, it was intended
to be page addressing mode which is 0x02. By accident the overflow behaviour
of horizontal addressing mode matches the algorithm u8g2 uses to update the
display, thus it still worked.
This commit changes all SSD1306-like displays to be actually setup in page
addressing mode.